+++ test/datalevin/search_test.clj
@@ -40,6 +40,10 @@
         engine ^SearchEngine (sut/new-search-engine lmdb)]
     (add-docs sut/add-doc engine)
 
+    (is (sut/doc-indexed? engine :doc4))
+    (is (not (sut/doc-indexed? engine :non-existent)))
+    (is (not (sut/doc-indexed? engine "non-existent")))
+
     (let [[tid mw ^SparseIntArrayList sl]
           (l/get-value lmdb c/terms "red" :string :term-info true)]
       (is (= (l/range-count lmdb c/terms [:all] :string) 32))
@@ -81,6 +85,7 @@
       (sut/remove-doc engine :doc1)
       (let [[tid mw ^SparseIntArrayList sl]
             (l/get-value lmdb c/terms "red" :string :term-info true)]
+        (is (not (sut/doc-indexed? engine :doc1)))
         (is (= (l/range-count lmdb c/docs [:all]) 4))
         (is (not (sl/contains-index? sl 1)))
         (is (= (sl/size sl) 3))
@@ -140,8 +145,13 @@
        [:put "raw" 3 "Moby Dick is a story of a whale and a man obsessed."]])
     (doseq [i [1 2 3]]
       (sut/add-doc engine i (l/get-value lmdb "raw" i)))
+    (is (not (sut/doc-indexed? engine 0)))
+    (is (sut/doc-indexed? engine 1))
     (is (= (sut/search engine "lazy") [1]))
     (is (= (sut/search engine "red" ) [1 2]))
     (is (= (sut/search engine "red" {:display :offsets})
            [[1 [["red" [10 39]]]] [2 [["red" [40]]]]]))
+    (testing "update"
+      (sut/add-doc engine 1 "The quick fox jumped over the lazy dogs.")
+      (is (= (sut/search engine "red" ) [2])))
     (l/close-kv lmdb)))